api/requirements.txt
Flask==3.0.0
waitress==2.1.2
python-dotenv==1.0.0
Pylint==3.0.2
pydub==0.25.1
pytest==7.4.3
pytest-asyncio==0.23.2
websockets==12.0
SpeechRecognition==3.10.1
